JK builder refactoring and optimization (#505)
* Add cu files

* slight optiimzation

* for RSH functional

* Debugging get_k

* Add build_k_o0

* Update logger

* Cleanup get_k

* slight optimization

* Unrolling rys-k

* Fix get_k initialization bug

* cleanup

* Reduce shm usage

* Improve rys_contract_jk

* Improve ejk_contract_ip1 functions

* Update rys_contract_jk_ip1

* rys_contract_jk_ip1 cleanup

* Update rys_contract_jk_ip2

* Update rys_contract_j

* Cleanup MD-J code

* Update int3c2e

* cleanup pbc head files

* Replace vhf1.cuh

* Update PBCIntEnvVars structure

* fix unrolled_rys_k

* lint

* Update pbc.cuh

* Fix hessian JK builder

* Fix get_k for density-fitting instances

* Fix JK builder for hf_lowmem and rks_lowmem

* import error

* Fix task buffer overlap when two blocks are assigned to one sm

* Fix shm overwritten issue in rys_contract_jk_ip1

---------

+++ gpu4pyscf/lib/gint-rys/int3c2e.cuh
@@ -31,23 +31,23 @@ typedef struct {
 } Int3c2eEnvVars;
 
 typedef struct {
-    uint8_t li;
-    uint8_t lj;
-    uint8_t lk;
-    uint8_t nroots;
-    uint8_t nfi;
-    uint8_t nfij;
-    uint8_t nfk;
-    uint8_t iprim;
-    uint8_t jprim;
-    uint8_t kprim;
-    uint8_t stride_i;
-    uint8_t stride_j;
-    uint8_t stride_k;
-    uint8_t g_size;
-    uint16_t naux;
-    uint16_t nksh;
-    uint16_t ksh0;
+    int li;
+    int lj;
+    int lk;
+    int nroots;
+    int nfi;
+    int nfij;
+    int nfk;
+    int iprim;
+    int jprim;
+    int kprim;
+    int stride_i;
+    int stride_j;
+    int stride_k;
+    int g_size;
+    int naux;
+    int nksh;
+    int ksh0;
     int nshl_pair;
     // The effective basis pair Id = ish*nbas+jsh
     int *bas_ij_idx;
